Make avr and sam files more similar to aid #5771
