+++ hardware/arduino/sam/cores/arduino/Print.cpp
@@ -41,11 +41,6 @@ size_t Print::write(const uint8_t *buffer, size_t size)
   return n;
 }
 
-size_t Print::print(const __FlashStringHelper *ifsh)
-{
-  return print(reinterpret_cast<const char *>(ifsh));
-}
-
 size_t Print::print(const String &s)
 {
   return write(s.c_str(), s.length());
